+++ BookWorm/src/main/java/com/bookworm/services/ProductAttributeService.java
@@ -9,9 +9,9 @@ public interface ProductAttributeService {
 
 
 	 List<ProductAttribute> getAllProductAttributes();
-	 Optional<ProductAttribute> getProductAttributeById(int id);
+	 Optional<ProductAttribute> getProductAttributeById(long id);
 	 ProductAttribute createProductAttribute(ProductAttribute productAttribute) ;
-	 void updateProductAttribute(int id, ProductAttribute updatedProductAttribute);
-	 void deleteProductAttribute(int id);
+	 void updateProductAttribute(long id, ProductAttribute updatedProductAttribute);
+	 void deleteProductAttribute(long id);
 
 }
